requirements.txt
fastapi
sqlalchemy
pymysql
uvicorn
python-dotenv
